Declared VEnum destructor
I'm not actually sure what this error was. AssetDock.cpp was showing a build error if ~VEnum() was defined in VEnum.cpp.
The other error was Material was throwing an error with its VEnum inside of it on cleanup. No idea what it was. Maybe something to do with std::map