Fix Polyboard buffer creation crash with calced byte aligned sizes
There was an issue where creating dynamic vertex and index buffers on Polyboard was crashing, and I think it was because the byte size specified wasn't aligned to multiples of 16. Previous todo here said it was about mapping the buffer after creation, but d3d11 would always hit an exception if pSysMem was left NULL. I think the alignment fixes the issue, but not sure, as the original crash didn't always happen anyway. It did leave some weird visual CRT static like artefacts in-world though sometimes.