Removed cached world matrix from Transform
Previously, local and world matrices were cached on a SpatialComponent's transform by calling UpdateTransform() on each of the inner transform sets within SpatialComponent. This change means everything is calling GetWorldMatrix() instead, so the tradeoff between UpdateComponent() being called on SetPosition() and Trasnform::world being cached to use on GetForwardVector() for example, basically evened each other out as far as it being an "optimisation".